+++ src/core/ui/card/ContributorCardHorizontal.tsx
@@ -1,26 +1,31 @@
+import React, { ReactNode, useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Box } from '@mui/material';
 import Avatar from '../avatar/Avatar';
-import { BlockSectionTitle } from '../typography';
 import getLocationString, { Location } from '../location/getLocationString';
 import BadgeCardView from '../list/BadgeCardView';
-import React, { useMemo } from 'react';
 import GridItem from '../grid/GridItem';
 import SwapColors from '../palette/SwapColors';
-import { useTranslation } from 'react-i18next';
-import ContributorTooltip from './ContributorTooltip';
 import RouterLink from '../link/RouterLink';
+import { BlockSectionTitle } from '../typography';
+import ContributorTooltip from './ContributorTooltip';
 
 export interface SpaceWelcomeSectionContributorProps {
-  profile: {
-    displayName: string;
-    avatar?: {
-      uri: string;
-    };
-    location?: Location;
-    tagsets?: { tags: string[] }[];
-  };
-  url: string;
+  profile:
+    | {
+        displayName: string;
+        avatar?: {
+          uri: string;
+        };
+        location?: Location;
+        tagsets?: { tags: string[] }[];
+      }
+    | undefined;
+  url?: string;
   onContact?: () => void;
   seamless?: boolean;
+  actions?: ReactNode;
+  index?: ReactNode;
 }
 
 const ContributorCardHorizontal = ({
@@ -28,12 +33,14 @@ const ContributorCardHorizontal = ({
   profile,
   onContact,
   seamless = false,
+  actions,
+  index,
 }: SpaceWelcomeSectionContributorProps) => {
   const { t } = useTranslation();
 
   const tags = useMemo(
-    () => (profile.tagsets ? profile.tagsets.flatMap(tagset => tagset.tags) : []),
-    [profile.tagsets]
+    () => (profile?.tagsets ? profile.tagsets.flatMap(tagset => tagset.tags) : []),
+    [profile?.tagsets]
   );
 
   if (!seamless) {
@@ -45,34 +52,42 @@ const ContributorCardHorizontal = ({
   }
 
   return (
-    <GridItem columns={2}>
-      <ContributorTooltip
-        displayName={profile.displayName}
-        avatarSrc={profile.avatar?.uri}
-        tags={tags}
-        city={profile.location?.city}
-        country={profile.location?.country}
-        isContactable={!!onContact}
-        onContact={onContact}
-      >
-        <BadgeCardView
-          visual={
-            <Avatar
-              src={profile.avatar?.uri}
-              aria-label="User avatar"
-              alt={t('common.avatar-of', { user: profile.displayName })}
-            >
-              {profile.displayName[0]}
-            </Avatar>
-          }
-          component={RouterLink}
-          to={url}
+    <>
+      <GridItem columns={2}>
+        <ContributorTooltip
+          displayName={profile?.displayName}
+          avatarSrc={profile?.avatar?.uri}
+          tags={tags}
+          city={profile?.location?.city}
+          country={profile?.location?.country}
+          isContactable={!!onContact}
+          onContact={onContact}
         >
-          <BlockSectionTitle>{profile.displayName}</BlockSectionTitle>
-          <BlockSectionTitle>{profile.location && getLocationString(profile.location)}</BlockSectionTitle>
-        </BadgeCardView>
-      </ContributorTooltip>
-    </GridItem>
+          <BadgeCardView
+            visual={
+              <Avatar
+                src={profile?.avatar?.uri}
+                aria-label="User avatar"
+                alt={t('common.avatar-of', { user: profile?.displayName })}
+              >
+                {profile?.displayName?.[0]}
+              </Avatar>
+            }
+            component={RouterLink}
+            to={url ?? ''}
+          >
+            <Box display="flex" flexDirection="row" justifyContent="space-between">
+              <Box display="flex" flexDirection="column">
+                <BlockSectionTitle>{profile?.displayName}</BlockSectionTitle>
+                <BlockSectionTitle>{profile?.location && getLocationString(profile.location)}</BlockSectionTitle>
+              </Box>
+              {index}
+            </Box>
+          </BadgeCardView>
+        </ContributorTooltip>
+      </GridItem>
+      {actions}
+    </>
   );
 };
